add recipe-ingredient and review fixes

+++ cookbookapp/resources/recipe_ingredient_qty.py
@@ -1,43 +1,27 @@
+"""
+This module contains the resources for handling recipe-ingredient related API endpoints.
+"""
 import json
 import logging
 from flask_restful import Resource
 from flask import Response, request, url_for
 from jsonschema import ValidationError, validate
-from cookbookapp import db
 from sqlalchemy.exc import IntegrityError
+from cookbookapp import db
 from cookbookapp.models import RecipeIngredientQty
 
 logging.basicConfig(level=logging.INFO)
 
-class IngredientQtyCollection(Resource):
-    def get(self):
-        body = {"items": []}
-        # body["self_uri"] = url_for("api.ingredientqtycollection")
-        # body["name"] = "Recipe Ingredient Qty Collection"
-        # body["description"] = "A collection of recipe ingredient quantities"
-
-        # body["controls"] = {
-        #     "create_recipe_ingredient_qty": {"method": "POST", "href": url_for("api.ingredientqtycollection"), "title": "Create a new recipe ingredient quantity", "schema": RecipeIngredientQty.get_schema()},
-        #     "search_recipe_ingredient_qty": {"method": "GET", "href": "/api/ingredientqty/search", "title": "Search for recipes"} # not implemented
-        # }
-
-        ingreQtys = RecipeIngredientQty.query.all()
-        for ingreQty in ingreQtys:
-            
-            item = ingreQty.serialize()
-            # item["controls"] = {
-            #     "self": {"method": "GET", "href": url_for("api.ingredientqtyitem", ingredientqty=ingredientqty.qty_id), "title": "Recipe Ingredient Quantity details"},
-            #     "update": {"method": "PUT", "href": url_for("api.ingredientqtyitem", ingredientqty=ingredientqty.qty_id), "title": "Update recipe ingredient quantity", "schema": RecipeIngredientQty.get_schema()}, # for reference
-            #     "delete": {"method": "DELETE", "href": url_for("api.ingredientqtyitem", ingredientqty=ingredientqty.qty_id), "title": "Delete recipe ingredient quantity"} # for reference
-            # }
-
-            body["items"].append(item)
-
-        return Response(json.dumps(body), status=200, mimetype="application/json")
-
-    def post(self):
+class RecipeIngredientQtyCollection(Resource):
+    """
+    Represents a collection of recipe-ingredients.
+    """
+    def post(self, recipe):
+        """
+        Handle POST requests to create a new recipe-ingredient.
+        """
         if not request.is_json:
-            body = { 
+            body = {
                 "error": {
                     "title": "Unsupported media type",
                     "description": "Requests must be JSON"
@@ -57,10 +41,10 @@ class IngredientQtyCollection(Resource):
             return Response(json.dumps(body), status=400, mimetype="application/json")
 
         ingredientqty = RecipeIngredientQty(
-            recipe_id=request.json["recipe_id"],
+            recipe_id=recipe.recipe_id,
             ingredient_id=request.json["ingredient_id"],
             qty=request.json["qty"],
-            metric=request.json["metric"]
+            metric=request.get_json().get("metric", "g")
         )
 
         try:
@@ -74,30 +58,28 @@ class IngredientQtyCollection(Resource):
                 }
             }
             return Response(json.dumps(body), status=409, mimetype="application/json")
-        
-        return Response(status=201, headers={
-            "Location": url_for("api.ingredientqtyitem", ingredientqty=ingredientqty)
-        })
-    
-class IngredientQtyItem(Resource):
-    def get(self, ingredientqty):
-        body = ingredientqty.serialize()
-        # body["controls"] = {
-        #     "ingredientqty:update": {"method": "PUT", "href": url_for("api.ingredientqtyitem", ingredientqty=ingredientqty.qty_id), "title": "Update recipe ingredient quantity", "schema": RecipeIngredientQty.get_schema()},
-        #     "ingredientqty:delete": {"method": "DELETE", "href": url_for("api.ingredientqtyitem", ingredientqty=ingredientqty.qty_id), "title": "Delete recipe ingredient quantity"}
-        # }
-        return Response(json.dumps(body), status=200, mimetype="application/json")
 
-    def put(self, ingredientqty):
+        return Response(status=201, headers={
+            "Location": url_for("api.recipeingredientqtyitem", ingredientqty=ingredientqty)
+        })
+
+class RecipeIngredientQtyItem(Resource):
+    """
+    Represents a single recipe ingredient.
+    """
+    def put(self, recipe, ingredient):
+        """
+        Handle GET requests to retrieve a single recipe ingredient.
+        """
         if not request.is_json:
-            body = { 
+            body = {
                 "error": {
                     "title": "Unsupported media type",
                     "description": "Requests must be JSON"
                 }
             }
             return Response(json.dumps(body), status=415, mimetype="application/json")
-    
+
         try:
             validate(request.json, RecipeIngredientQty.get_schema())
         except ValidationError as e:
@@ -108,10 +90,16 @@ class IngredientQtyItem(Resource):
                 }
             }
             return Response(json.dumps(body), status=400, mimetype="application/json")
-        
-        #ingredientqty.qty_id = request.json["qty_id"]
-        ingredientqty.recipe_id = request.json["recipe_id"]
-        ingredientqty.ingredient_id = request.json["ingredient_id"]
+        ingredientqty = RecipeIngredientQty.query.filter_by(
+            recipe_id=recipe.recipe_id ,ingredient_id=ingredient.ingredient_id).first()
+        if not ingredientqty:
+            body = {
+                "error": {
+                    "title": "Not Found",
+                    "description": "Recipe Ingredient Quantity not found"
+                }
+            }
+            return Response(json.dumps(body), status=404, mimetype="application/json")
         ingredientqty.qty = request.json["qty"]
         ingredientqty.metric = request.json["metric"]
 
@@ -125,11 +113,23 @@ class IngredientQtyItem(Resource):
                 }
             }
             return Response(json.dumps(body), status=409, mimetype="application/json")
-        
+
         return Response(status=204)
-    
-    def delete(self, ingredientqty):
+
+    def delete(self, recipe, ingredient):
+        """
+        Handle DELETE requests to delete a recipe ingredient.
+        """
+        ingredientqty = RecipeIngredientQty.query.filter_by(
+            recipe_id=recipe.recipe_id ,ingredient_id=ingredient.ingredient_id).first()
+        if not ingredientqty:
+            body = {
+                "error": {
+                    "title": "Not Found",
+                    "description": "Recipe Ingredient Quantity not found"
+                }
+            }
+            return Response(json.dumps(body), status=404, mimetype="application/json")
         db.session.delete(ingredientqty)
         db.session.commit()
         return {"message": "Recipe Ingredient Qty deleted"}, 204
-
